refactor(ReservationCounter): 마이너스 버튼, 플러스 버튼, 카운터 텍스트에 대한 뷰들을 하나로 관리하는 클래스 추가
